Tests/LibWeb/Text/expected/css/embedded-about-blank-color-scheme.txt

Parent root color-scheme: normal
Initial iframe canvas color-scheme: light
Navigated iframe canvas color-scheme: light
